Removed Spielfeld class because of wrong use of Beet[][] object
