+++ .settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+encoding//src/de/dhstudent/fruitfield/Beet.java=UTF-8
